requirements.txt
telethon==1.36.0
anthropic>=0.42.0
google-api-python-client==2.154.0
google-auth==2.36.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
